requirements.txt
request==2019.4.13
requests_oauthlib==1.3.0
oauthlib==3.1.0